Fix SDL tiles arena UI bugs
Fixes the arena crashing and otherwise failing to handle window resizing
correctly.

A mostly straight-forward conversion. ui_delay() is a replacement for
delay() that still responds to UI events, although the delay currently
cannot be interrupted (which is sub-optimal).

In tilereg-map.cc, add_square() checks the current transform to ensure
that the square is pixel perfect regardless of the transform, so we need
to match the transforms used at buffer-packing- and at render-time.