Add item/spell/ability menu adjust mode
Functionality is implemented in menu.cc, and is enabled whenever a menu
has an `on_entry_swap` callback, which does the actual swapping. An
additional `swap_mode_prompt_noun` member is used to generate the swap
prompt text. The on_entry_swap callback receives the first item to swap,
and either a pointer to the second or a key to swap it to.

Swap mode is entered by pressing = on a menu that supports it. An
already-existing item is then selected, followed by a letter that may or
may not refer to an existing item. Swapping an item to its own letter
cancels the swap, but stays in swap mode. At any time, = or ESC cancels
the swap and leaves swap mode.