Show msgwin_get_line as a popup when necessary
i.e. when there is no game currently playing, or when one or more UI
layouts are already being shown (and the message region is obscured).

This includes a new HTML input UI, as the problem of the message region
being obscured also existed on webtiles. Currently, this is very
simplistic, and doesn't use a text box; instead, keystrokes are
forwarded to the server, which then updates the HTML API; so every
keypress requires a round trip to render.

There's quite a few call sites that use cancellable_get_line directly;
it's unclear whether there's any special reason for doing so, but
they'll need to be converted to use msgwin_get_line() (if possible) to
reap the benefits of this commit.